+++ allauth/socialaccount/forms.py
@@ -33,7 +33,7 @@ class SignupForm(BaseSignupForm):
             return super(SignupForm, self).validate_unique_email(value)
         except forms.ValidationError:
             raise get_adapter().validation_error(
-                "email_taken", self.sociallogin.account.get_provider().name
+                "email_taken", self.sociallogin.provider.name
             )